A more extendable version
Hard code the resources we want to gather, but make the gathering
function generic enough so that we can easily add more. This means we're
collecting more data then before, but the idea would be that puppetdb
only stores the bits it needs.

This example adds users to show how easy this is to extend.