(PE-XXX) Gather package inventory
Here we add the ability to gather package inventory information and add
it to the report.

This functionality defaults to being disabled, and can be enabled by
setting the `pe_enabled` configuration option to `true`. When enabled
Puppet::Indirector.search will be used to gather a list of all installed
packages, modify the name value to be easier consumed by PDB and then
added to the report.

This solution is intended to be fairly generic since the SC&R team will
eventually need to gather more resources as this feature grows.